refactor(tools): split src/tools.py and src/tools_web.py into src/tools package
- Created src/tools/__init__.py with public re-exports and side-effect imports
- Created src/tools/registry.py with ToolDef, ToolRegistry, tool, ToolContext, ToolExecutor, default_registry
- Created src/tools/slack.py with read_attached_images, read_attached_document, fetch_thread_history
- Created src/tools/search.py with search_web, _ddg_search, _tavily_search
- Created src/tools/web.py merging fetch_webpage from tools.py + all SSRF/HTML/fetch helpers from tools_web.py
- Created src/tools/image.py with generate_image
- Created src/tools/time.py with get_current_time
- Deleted src/tools.py and src/tools_web.py
- Updated tests/test_tools.py import paths and patch targets to new submodule locations
