+++ src/tools/slack.py
@@ -1,167 +1,24 @@
-"""Tool registry + 6 built-in tools with JSON Schema specs.
-
-Tools are declared once via the `@tool(...)` decorator. The same registry
-produces JSON Schemas for LLM function calling AND the executor's dispatch
-table. A per-call timeout guards against slow network I/O from one tool
-blocking the whole agent loop.
-"""
+"""Slack-centric tools: read images/documents attached to the triggering
+mention, fetch the current thread's history."""
 from __future__ import annotations
 
-import json
 import logging
-import socket
 import time
 import urllib.error
 import urllib.parse
 import urllib.request
-from concurrent.futures import ThreadPoolExecutor, TimeoutError as FuturesTimeout
-from dataclasses import dataclass, field
 from typing import Any, Callable
 
-from botocore.exceptions import BotoCoreError, ClientError
 from slack_sdk.errors import SlackApiError
 
-from src.config import Settings
-from src.llms import LLMProvider, ToolCall
 from src.slack_helpers import user_name_cache
-
-# Re-export web-fetch helpers so callers (and tests that still import from
-# src.tools) keep working while the web-fetch module lives in tools_web.py.
-from src.tools_web import (
-    _HtmlTextExtractor,
-    _JINA_HEADER_MAX_LINES,
-    _JINA_LINK_RE,
-    _NoRedirectHandler,
-    _PUBLIC_WEB_UA,
-    _WEB_FETCH_TIMEOUT,
-    _extract_markdown_links,
-    _filter_links,
-    _jina_fetch,
-    _parse_jina_response,
-    _raw_fetch,
-    _read_body_capped,
-    _validate_public_https_url,
-)
+from src.tools.registry import ToolContext, default_registry, tool
 
 logger = logging.getLogger(__name__)
 
-
 SLACK_FILE_HOSTS = {"files.slack.com", "files-edge.slack.com", "files-pri.slack.com"}
-DUCKDUCKGO_HOST = "api.duckduckgo.com"
-TAVILY_HOST = "api.tavily.com"
 DOC_TEXT_PREFIX = "text/"
 DOC_PDF_MIME = "application/pdf"
-
-
-# --------------------------------------------------------------------------- #
-# Registry
-# --------------------------------------------------------------------------- #
-
-
-@dataclass
-class ToolDef:
-    name: str
-    description: str
-    parameters: dict[str, Any]  # JSON Schema
-    fn: Callable[..., Any]
-    timeout: float | None = None  # None -> use executor default
-
-
-@dataclass
-class ToolRegistry:
-    _tools: dict[str, ToolDef] = field(default_factory=dict)
-
-    def register(self, td: ToolDef) -> None:
-        self._tools[td.name] = td
-
-    def names(self) -> list[str]:
-        return list(self._tools.keys())
-
-    def get(self, name: str) -> ToolDef | None:
-        return self._tools.get(name)
-
-    def specs(self) -> list[dict[str, Any]]:
-        return [
-            {"name": t.name, "description": t.description, "parameters": t.parameters}
-            for t in self._tools.values()
-        ]
-
-
-def tool(
-    registry: ToolRegistry,
-    name: str,
-    description: str,
-    parameters: dict[str, Any],
-    timeout: float | None = None,
-):
-    def decorator(fn: Callable[..., Any]):
-        registry.register(
-            ToolDef(name=name, description=description, parameters=parameters, fn=fn, timeout=timeout)
-        )
-        return fn
-
-    return decorator
-
-
-# --------------------------------------------------------------------------- #
-# Context
-# --------------------------------------------------------------------------- #
-
-
-@dataclass
-class ToolContext:
-    slack_client: Any
-    channel: str
-    thread_ts: str
-    event: dict[str, Any]
-    settings: Settings
-    llm: LLMProvider
-
-
-# --------------------------------------------------------------------------- #
-# Executor
-# --------------------------------------------------------------------------- #
-
-
-class ToolExecutor:
-    def __init__(self, context: ToolContext, registry: ToolRegistry, timeout: float = 20.0):
-        self.context = context
-        self.registry = registry
-        self.timeout = timeout
-        self._pool = ThreadPoolExecutor(max_workers=2)
-
-    def execute(self, call: ToolCall) -> dict[str, Any]:
-        td = self.registry.get(call.name)
-        started = time.monotonic()
-        if td is None:
-            return {"ok": False, "error": f"unknown tool: {call.name}"}
-        effective_timeout = td.timeout if td.timeout is not None else self.timeout
-        try:
-            future = self._pool.submit(td.fn, self.context, **(call.arguments or {}))
-            result = future.result(timeout=effective_timeout)
-            return {"ok": True, "result": result, "duration_ms": int((time.monotonic() - started) * 1000)}
-        except FuturesTimeout:
-            logger.warning("tool %s timed out after %.1fs", call.name, effective_timeout)
-            return {"ok": False, "error": f"tool '{call.name}' timed out after {effective_timeout}s"}
-        except (
-            TypeError,
-            ValueError,
-            KeyError,
-            urllib.error.URLError,
-            json.JSONDecodeError,
-            SlackApiError,
-            BotoCoreError,
-            ClientError,
-        ) as exc:
-            logger.exception("tool %s failed", call.name)
-            return {"ok": False, "error": f"{exc.__class__.__name__}: {exc}"}
-
-
-# --------------------------------------------------------------------------- #
-# Built-in tools
-# --------------------------------------------------------------------------- #
-
-default_registry = ToolRegistry()
 
 
 @tool(
@@ -494,209 +351,6 @@ def fetch_thread_history(ctx: ToolContext, limit: int = 20) -> list[dict[str, An
     )
 
 
-@tool(
-    default_registry,
-    name="search_web",
-    description="Search the public web for up-to-date information. Uses Tavily if TAVILY_API_KEY is set, otherwise DuckDuckGo Instant Answer.",
-    parameters={
-        "type": "object",
-        "properties": {
-            "query": {"type": "string"},
-            "limit": {"type": "integer", "minimum": 1, "maximum": 20, "default": 5},
-        },
-        "required": ["query"],
-    },
-)
-def search_web(ctx: ToolContext, query: str, limit: int = 5) -> list[dict[str, str]]:
-    if ctx.settings.tavily_api_key:
-        return _tavily_search(ctx.settings.tavily_api_key, query, limit)
-    return _ddg_search(query, limit)
-
-
-@tool(
-    default_registry,
-    name="fetch_webpage",
-    description=(
-        "Fetch a public HTTPS web page and return clean text content plus a "
-        "list of outbound links. Use this for summarizing articles, landing "
-        "pages, or news indexes. For Slack-hosted files use "
-        "read_attached_document / read_attached_images instead."
-    ),
-    parameters={
-        "type": "object",
-        "properties": {
-            "url": {
-                "type": "string",
-                "description": "Absolute https URL of the page to fetch.",
-            },
-            "max_chars": {
-                "type": "integer",
-                "minimum": 500,
-                "maximum": 20000,
-                "description": "Optional cap on returned content characters. Clamped by MAX_WEB_CHARS.",
-            },
-            "max_links": {
-                "type": "integer",
-                "minimum": 0,
-                "maximum": 50,
-                "description": "Optional cap on returned link count. Clamped by MAX_WEB_LINKS.",
-            },
-        },
-        "required": ["url"],
-    },
-    timeout=30.0,
-)
-def fetch_webpage(
-    ctx: ToolContext,
-    url: str,
-    max_chars: int | None = None,
-    max_links: int | None = None,
-) -> dict[str, Any]:
-    _validate_public_https_url(url)
-    settings = ctx.settings
-    effective_chars = min(
-        max_chars if max_chars is not None else settings.max_web_chars,
-        settings.max_web_chars,
-    )
-    effective_links = min(
-        max_links if max_links is not None else settings.max_web_links,
-        settings.max_web_links,
-    )
-
-    title = ""
-    content = ""
-    links: list[dict[str, str]] = []
-    jina_ok = False
-    jina_err: str | None = None
-
-    # HTTPError is a subclass of URLError; we list both to make the
-    # "HTTP status error" path visible as a first-class error type for
-    # readers. Order doesn't affect matching (URLError catches both).
-    try:
-        jina_text = _jina_fetch(settings.jina_reader_base, url, settings.max_web_bytes)
-        parsed_title, body = _parse_jina_response(jina_text)
-        if body.strip():
-            title = parsed_title
-            content = body
-            links = _extract_markdown_links(body, url, effective_links)
-            jina_ok = True
-        else:
-            jina_err = "empty body"
-    except (
-        ValueError,
-        urllib.error.HTTPError,
-        urllib.error.URLError,
-        socket.timeout,
-        UnicodeDecodeError,
-    ) as exc:
-        jina_err = f"{exc.__class__.__name__}: {exc}"
-        logger.debug("fetch_webpage jina fallback (url=%s): %s", url, jina_err)
-
-    if not jina_ok:
-        # Belt-and-suspenders: re-validate before the raw fetch. _raw_fetch
-        # assumes its caller has passed the URL through
-        # _validate_public_https_url; re-checking here keeps that contract
-        # enforced even if future refactors mutate `url` in between.
-        _validate_public_https_url(url)
-        try:
-            html_text = _raw_fetch(url, settings.max_web_bytes)
-        except (
-            ValueError,
-            urllib.error.HTTPError,
-            urllib.error.URLError,
-            socket.timeout,
-        ) as exc:
-            raise ValueError(
-                f"fetch_webpage failed: jina={jina_err}, raw={exc.__class__.__name__}: {exc}"
-            ) from exc
-        extractor = _HtmlTextExtractor(url)
-        extractor.feed(html_text)
-        title = extractor.title() or title
-        content = extractor.text()
-        links = _filter_links(extractor.links, url, effective_links)
-
-    source = "jina" if jina_ok else "raw"
-
-    truncated = False
-    if len(content) > effective_chars:
-        content = content[:effective_chars]
-        truncated = True
-
-    return {
-        "url": url,
-        "title": title,
-        "content": content,
-        "links": links,
-        "chars": len(content),
-        "truncated": truncated,
-        "source": source,
-    }
-
-
-@tool(
-    default_registry,
-    name="generate_image",
-    description="Generate an image from a prompt and upload it to the Slack thread. Returns the permalink.",
-    parameters={
-        "type": "object",
-        "properties": {"prompt": {"type": "string"}},
-        "required": ["prompt"],
-    },
-    timeout=75.0,  # gpt-image-1 / titan / stability can take 30–60s
-)
-def generate_image(ctx: ToolContext, prompt: str) -> dict[str, str]:
-    image_bytes = ctx.llm.generate_image(prompt)
-    upload = ctx.slack_client.files_upload_v2(
-        channel=ctx.channel,
-        thread_ts=ctx.thread_ts,
-        title="Generated image",
-        filename="generated.png",
-        file=image_bytes,
-    )
-    file_info = upload.get("file", {})
-    return {"permalink": file_info.get("permalink", ""), "title": file_info.get("title", "generated.png")}
-
-
-@tool(
-    default_registry,
-    name="get_current_time",
-    description=(
-        "Return the current wall-clock time. Uses the server default "
-        "timezone (DEFAULT_TIMEZONE env) unless 'timezone' is provided. "
-        "Useful for 'today', 'now', 'this week', or weekday questions."
-    ),
-    parameters={
-        "type": "object",
-        "properties": {
-            "timezone": {
-                "type": "string",
-                "description": (
-                    "Optional IANA timezone (e.g. 'Asia/Seoul', 'UTC', "
-                    "'America/New_York'). Omit to use the server default."
-                ),
-            }
-        },
-        "required": [],
-    },
-)
-def get_current_time(ctx: ToolContext, timezone: str | None = None) -> dict[str, Any]:
-    from datetime import datetime
-    from zoneinfo import ZoneInfo, ZoneInfoNotFoundError
-
-    tz_name = timezone or ctx.settings.default_timezone
-    try:
-        tz = ZoneInfo(tz_name)
-    except ZoneInfoNotFoundError as exc:
-        raise ValueError(f"unknown timezone: {tz_name}") from exc
-    now = datetime.now(tz)
-    return {
-        "iso": now.isoformat(timespec="seconds"),
-        "timezone": tz_name,
-        "weekday": now.strftime("%A"),
-        "unix": int(now.timestamp()),
-    }
-
-
 # --------------------------------------------------------------------------- #
 # Helpers
 # --------------------------------------------------------------------------- #
@@ -721,40 +375,3 @@ def _with_slack_retry(call: Callable[[], Any], map_result: Callable[[Any], Any],
     if last:
         raise last
     return []
-
-
-def _ddg_search(query: str, limit: int) -> list[dict[str, str]]:
-    params = urllib.parse.urlencode({"q": query, "format": "json", "no_redirect": 1, "no_html": 1})
-    url = f"https://{DUCKDUCKGO_HOST}/?{params}"
-    parsed = urllib.parse.urlparse(url)
-    if parsed.scheme != "https" or parsed.hostname != DUCKDUCKGO_HOST:
-        raise ValueError("invalid web search URL")
-    with urllib.request.urlopen(url, timeout=15) as response:  # noqa: S310
-        payload = json.loads(response.read().decode("utf-8"))
-    results: list[dict[str, str]] = []
-    if payload.get("AbstractURL"):
-        results.append({"title": payload.get("AbstractText", ""), "url": payload["AbstractURL"]})
-    for item in payload.get("RelatedTopics", []):
-        if "Text" in item and "FirstURL" in item:
-            results.append({"title": item["Text"], "url": item["FirstURL"]})
-            if len(results) >= limit:
-                break
-    return results[:limit]
-
-
-def _tavily_search(api_key: str, query: str, limit: int) -> list[dict[str, str]]:
-    url = f"https://{TAVILY_HOST}/search"
-    parsed = urllib.parse.urlparse(url)
-    if parsed.scheme != "https" or parsed.hostname != TAVILY_HOST:
-        raise ValueError("invalid Tavily URL")
-    body = json.dumps({"api_key": api_key, "query": query, "max_results": limit}).encode()
-    req = urllib.request.Request(url, data=body, headers={"Content-Type": "application/json"}, method="POST")
-    with urllib.request.urlopen(req, timeout=15) as response:  # noqa: S310
-        payload = json.loads(response.read().decode("utf-8"))
-    return [
-        {"title": r.get("title", ""), "url": r.get("url", ""), "content": r.get("content", "")}
-        for r in payload.get("results", [])[:limit]
-    ]
-
-
-
