+++ src/tools/web.py
@@ -1,18 +1,24 @@
-"""Web-fetch helpers used by the fetch_webpage tool.
+"""Public-HTTPS web fetch (Jina Reader with raw-urllib fallback) + SSRF guard.
 
-Split out from src/tools.py to keep each module focused. Symbols defined
-here are re-exported from src.tools for backward compatibility with test
-imports and callers.
+All helpers live here so the SSRF pre-flight, the no-redirect handler, the
+Jina/raw fetchers, and the HTML/markdown parsers can be audited together.
+Test patches target this module's urlopen / build_opener / getaddrinfo names.
 """
 from __future__ import annotations
 
 import ipaddress
+import logging
 import re
 import socket
 import urllib.error
 import urllib.parse
 import urllib.request
 from html.parser import HTMLParser
+from typing import Any
+
+from src.tools.registry import ToolContext, default_registry, tool
+
+logger = logging.getLogger(__name__)
 
 
 _PUBLIC_WEB_UA = "lambda-gurumi-bot/1.0 (+https://github.com/nalbam/lambda-gurumi-bot)"
@@ -268,3 +274,123 @@ def _parse_jina_response(text: str) -> tuple[str, str]:
         return title, text
     body = "\n".join(all_lines[body_start:]).lstrip("\n")
     return title, body
+
+
+@tool(
+    default_registry,
+    name="fetch_webpage",
+    description=(
+        "Fetch a public HTTPS web page and return clean text content plus a "
+        "list of outbound links. Use this for summarizing articles, landing "
+        "pages, or news indexes. For Slack-hosted files use "
+        "read_attached_document / read_attached_images instead."
+    ),
+    parameters={
+        "type": "object",
+        "properties": {
+            "url": {
+                "type": "string",
+                "description": "Absolute https URL of the page to fetch.",
+            },
+            "max_chars": {
+                "type": "integer",
+                "minimum": 500,
+                "maximum": 20000,
+                "description": "Optional cap on returned content characters. Clamped by MAX_WEB_CHARS.",
+            },
+            "max_links": {
+                "type": "integer",
+                "minimum": 0,
+                "maximum": 50,
+                "description": "Optional cap on returned link count. Clamped by MAX_WEB_LINKS.",
+            },
+        },
+        "required": ["url"],
+    },
+    timeout=30.0,
+)
+def fetch_webpage(
+    ctx: ToolContext,
+    url: str,
+    max_chars: int | None = None,
+    max_links: int | None = None,
+) -> dict[str, Any]:
+    _validate_public_https_url(url)
+    settings = ctx.settings
+    effective_chars = min(
+        max_chars if max_chars is not None else settings.max_web_chars,
+        settings.max_web_chars,
+    )
+    effective_links = min(
+        max_links if max_links is not None else settings.max_web_links,
+        settings.max_web_links,
+    )
+
+    title = ""
+    content = ""
+    links: list[dict[str, str]] = []
+    jina_ok = False
+    jina_err: str | None = None
+
+    # HTTPError is a subclass of URLError; we list both to make the
+    # "HTTP status error" path visible as a first-class error type for
+    # readers. Order doesn't affect matching (URLError catches both).
+    try:
+        jina_text = _jina_fetch(settings.jina_reader_base, url, settings.max_web_bytes)
+        parsed_title, body = _parse_jina_response(jina_text)
+        if body.strip():
+            title = parsed_title
+            content = body
+            links = _extract_markdown_links(body, url, effective_links)
+            jina_ok = True
+        else:
+            jina_err = "empty body"
+    except (
+        ValueError,
+        urllib.error.HTTPError,
+        urllib.error.URLError,
+        socket.timeout,
+        UnicodeDecodeError,
+    ) as exc:
+        jina_err = f"{exc.__class__.__name__}: {exc}"
+        logger.debug("fetch_webpage jina fallback (url=%s): %s", url, jina_err)
+
+    if not jina_ok:
+        # Belt-and-suspenders: re-validate before the raw fetch. _raw_fetch
+        # assumes its caller has passed the URL through
+        # _validate_public_https_url; re-checking here keeps that contract
+        # enforced even if future refactors mutate `url` in between.
+        _validate_public_https_url(url)
+        try:
+            html_text = _raw_fetch(url, settings.max_web_bytes)
+        except (
+            ValueError,
+            urllib.error.HTTPError,
+            urllib.error.URLError,
+            socket.timeout,
+        ) as exc:
+            raise ValueError(
+                f"fetch_webpage failed: jina={jina_err}, raw={exc.__class__.__name__}: {exc}"
+            ) from exc
+        extractor = _HtmlTextExtractor(url)
+        extractor.feed(html_text)
+        title = extractor.title() or title
+        content = extractor.text()
+        links = _filter_links(extractor.links, url, effective_links)
+
+    source = "jina" if jina_ok else "raw"
+
+    truncated = False
+    if len(content) > effective_chars:
+        content = content[:effective_chars]
+        truncated = True
+
+    return {
+        "url": url,
+        "title": title,
+        "content": content,
+        "links": links,
+        "chars": len(content),
+        "truncated": truncated,
+        "source": source,
+    }
